IGNITE-22914 / switch to CRC-64

+++ modules/metastorage/src/main/java/org/apache/ignite/internal/metastorage/server/KeyValueStorage.java
@@ -381,5 +381,5 @@ public interface KeyValueStorage extends ManuallyCloseable {
      *
      * @param revision Revision.
      */
-    @Nullable Integer checksum(long revision);
+    @Nullable Long checksum(long revision);
 }
